refactor: moved VariationalFusionHead out of the stsnet scope (renamed to VariationalClassifier)
