+++ src/eegproc/supervised/__init__.py
@@ -1,0 +1,5 @@
+"""Supervised learning models and utilities for EEG emotion recognition."""
+
+from .variational_classifier import VariationalClassifier
+
+__all__ = ["VariationalClassifier"]
